Move emacs keybindings to eponymous file
