+++ src/edit_mode/mod.rs
@@ -4,6 +4,6 @@ mod keybindings;
 mod vi;
 
 pub use base::EditMode;
-pub use emacs::Emacs;
-pub use keybindings::{default_emacs_keybindings, Keybindings};
+pub use emacs::{default_emacs_keybindings, Emacs};
+pub use keybindings::Keybindings;
 pub use vi::{default_vi_insert_keybindings, default_vi_normal_keybindings, Vi};
